New mojo for generating Eclipse IP related information.
Refactor utility out of the existing mojos for re-use in other mojos and
add a new mojo for collecting source artifacts as well as refreshing
license information in about files.

This also allows to select a specific license for a dependency from a
list of known licenses in order to support electing a single, good
license from dual licensed bundles.

Additionally, it generates an ip_log.xml and allows to automatically
create CQs. The license is pre-selected from information in the
dependency pom. If no suitable license is found, a warning will be
logged and no CQ can be generated. In order to submit CQs, login
credentials need to be configured for a server with id
"portal.eclipse.org" in the local Maven settings.xml (password
encryption is recommended).

Note, the ip_log.xml license generation is not suitable for general
applicable for EBR recipes. It's more useful for Eclipse Orbit.

Change-Id: I238f856dce682683742c6d904bfc5cd61daf2283