Dependency declarations in osgi.bnd should have an upper bound.
A dependency declaration with no upper bound can become a problem when
newer versions of the dependency are introduced and the unbounded
requirement becomes forgotten.

Assuming some users may not be familiar with the format to specify a
"range", having them present by default gives some boilerplate and may
prevent conflicts.

Signed-off-by: Roland Grunberg <rgrunber@redhat.com>